python-consumer-kafka/requirements.txt
certifi==2023.5.7
confluent-kafka==2.1.1
console==0.9907
ezenv==0.92
minio==7.1.15
python-dotenv==1.0.0
urllib3==2.0.3